rename d3 utils and add create scale
