+++ app/assets/javascripts/components/plots/bar_plot.js.jsx
@@ -1,3 +1,4 @@
+import { createScale } from './utils/d3_scale'
 import YAxis from './yaxis'
 
 var BarPlot = React.createClass({
