+++ app/assets/javascripts/components/plots/utils/d3_scale.js
@@ -1,3 +1,14 @@
+export const createScale = (domain, range) => {
+  let scale
+  if (domain[0] instanceof Date) {
+    scale = d3.time.scale()
+  } else {
+    scale = d3.scale.linear()
+  }
+
+  return scale.range(range).domain(domain)
+}
+
 const tickDecimalPlaces = (ticks) => {
   const interval = Math.abs(ticks[1] - ticks[0])
   const places = Math.ceil(-Math.log10(interval))
@@ -16,6 +27,3 @@ export const tickFormatter = (ticks, scale) => {
 
   return (tick) => tick
 }
-
-
-
